requirements.txt
fastapi
uvicorn
groq
google-generativeai>=0.8.3
pydantic
python-multipart
pillow